+++ tests/TaskListTest.php
@@ -1,16 +1,23 @@
 <?php
- class CITest extends PHPUnit_Framework_TestCase
-  {
+use PHPUnit\Framework\TestCase;
+
+class TaskListTest extends TestCase
+{
     private $CI;
+
+    // Setup
     public function setUp()
     {
-      // Load CI instance normally
-      $this->CI = &get_instance();
+        $this->CI = &get_instance();
+        $this->CI->load->model('tasks');
+        $this->tasks = new Tasks();
     }
-    public function testGetPost()
+
+    // Check if task list is a valid size.
+    public function testListSize()
     {
-      $_SERVER['REQUEST_METHOD'] = 'GET';
-      $_GET['foo'] = 'bar';
-      $this->assertEquals('bar', $this->CI->input->get_post('foo'));
+        $completed = $this->tasks->getCompletedTask();
+        $uncompleted = $this->tasks->getUncompletedTask();
+        $this->assertLessThan($uncompleted, $completed);
     }
-  }
+}